Library/requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.1
